--- quote3/src/styles/avatar.module.css
@@ -1,11 +1,0 @@
-.avatarHolder {
-  border-radius: 50%;
-  background-color: #ccc; /* Default background color */
-  width: 37px;
-  aspect-ratio: 1/1;
-  overflow: hidden;
-}
-
-.avatar {
-  width: 100%;
-}
